Eliminamos la vista Publicar y la pasamos a muro, queda pendiente arreglar la promesa, porque aqui al parecer se rompe el codigo
